Move actor eval to behavior
